+++ app/page.tsx
@@ -3,7 +3,7 @@ import Trust from "@/components/home/Trust";
 import Services from "@/components/home/Services";
 import Process from "@/components/home/Process";
 import CaseHighlight from "@/components/home/CaseHighlight";
-import FAQ from "@/components/home/FAQ";
+import FAQSection from "@/components/home/FAQSection";
 import FinalCTA from "@/components/home/FinalCTA";
 import TestimonialsFull from "@/components/home/TestimonialsFull";
 
@@ -16,7 +16,7 @@ export default function Home() {
       <Process />
       <CaseHighlight />
       <TestimonialsFull />
-      <FAQ />
+      <FAQSection />
       <FinalCTA />
     </>
   );